+++ dsmr_backend/tests/test_services.py
@@ -30,12 +30,14 @@ class TestServices(InterceptStdoutMixin, TestCase):
         self.assertFalse(capabilities['electricity'])
         self.assertFalse(capabilities['electricity_returned'])
         self.assertFalse(capabilities['gas'])
+        self.assertFalse(capabilities['multi_phases'])
         self.assertFalse(capabilities['weather'])
         self.assertFalse(capabilities['any'])
 
         self.assertFalse(dsmr_backend.services.get_capabilities('electricity'))
         self.assertFalse(dsmr_backend.services.get_capabilities('electricity_returned'))
         self.assertFalse(dsmr_backend.services.get_capabilities('gas'))
+        self.assertFalse(dsmr_backend.services.get_capabilities('multi_phases'))
         self.assertFalse(dsmr_backend.services.get_capabilities('weather'))
         self.assertFalse(dsmr_backend.services.get_capabilities('any'))
 
@@ -57,6 +59,43 @@ class TestServices(InterceptStdoutMixin, TestCase):
         capabilities = dsmr_backend.services.get_capabilities()
         self.assertTrue(dsmr_backend.services.get_capabilities('electricity'))
         self.assertTrue(capabilities['electricity'])
+        self.assertTrue(capabilities['any'])
+
+    def test_multi_phases_capabilities(self):
+        """ Capability check for multiple phases. """
+        capabilities = dsmr_backend.services.get_capabilities()
+        self.assertFalse(capabilities['multi_phases'])
+        self.assertFalse(capabilities['any'])
+
+        ElectricityConsumption.objects.create(
+            read_at=timezone.now(),
+            delivered_1=0,
+            returned_1=0,
+            delivered_2=0,
+            returned_2=0,
+            currently_delivered=0,
+            currently_returned=0,
+            phase_currently_delivered_l1=1,
+        )
+
+        # Should fail.
+        self.assertFalse(dsmr_backend.services.get_capabilities('multi_phases'))
+
+        ElectricityConsumption.objects.create(
+            read_at=timezone.now() + timezone.timedelta(minutes=1),
+            delivered_1=0,
+            returned_1=0,
+            delivered_2=0,
+            returned_2=0,
+            currently_delivered=0,
+            currently_returned=0,
+            phase_currently_delivered_l2=1,
+            phase_currently_delivered_l3=1,
+        )
+        capabilities = dsmr_backend.services.get_capabilities()
+        self.assertTrue(dsmr_backend.services.get_capabilities('multi_phases'))
+
+        self.assertTrue(capabilities['multi_phases'])
         self.assertTrue(capabilities['any'])
 
     def test_electricity_returned_capabilities(self):