+++ dsmr_datalogger/tests/datalogger/test_landisgyr350_dsmr40.py
@@ -7,6 +7,7 @@ import pytz
 
 from dsmr_backend.tests.mixins import InterceptStdoutMixin
 from dsmr_datalogger.models.reading import DsmrReading, MeterStatistics
+from dsmr_datalogger.models.settings import DataloggerSettings
 
 
 class TestDatalogger(InterceptStdoutMixin, TestCase):
@@ -41,9 +42,9 @@ class TestDatalogger(InterceptStdoutMixin, TestCase):
             "1-0:31.7.0(000*A)\r\n",
             "1-0:51.7.0(000*A)\r\n",
             "1-0:71.7.0(001*A)\r\n",
-            "1-0:21.7.0(00.000*kW)\r\n",
-            "1-0:41.7.0(00.000*kW)\r\n",
-            "1-0:61.7.0(00.192*kW)\r\n",
+            "1-0:21.7.0(00.123*kW)\r\n",
+            "1-0:41.7.0(00.456*kW)\r\n",
+            "1-0:61.7.0(00.789*kW)\r\n",
             "1-0:22.7.0(00.000*kW)\r\n",
             "1-0:42.7.0(00.000*kW)\r\n",
             "1-0:62.7.0(00.000*kW)\r\n",
@@ -51,7 +52,7 @@ class TestDatalogger(InterceptStdoutMixin, TestCase):
             "0-1:96.1.0(xxxxxxxxxxxxx)\r\n",
             "0-1:24.2.1(151110190000W)(00845.206*m3)\r\n",
             "0-1:24.4.0(1)\r\n",
-            "!AF0C\n",
+            "!8CC9\n",
         ]
 
     @mock.patch('serial.Serial.open')
@@ -72,6 +73,9 @@ class TestDatalogger(InterceptStdoutMixin, TestCase):
 
     def test_reading_values(self):
         """ Test whether dsmr_datalogger reads the correct values. """
+        DataloggerSettings.get_solo()
+        DataloggerSettings.objects.all().update(track_phases=True)
+
         self._fake_dsmr_reading()
         self.assertTrue(DsmrReading.objects.exists())
         reading = DsmrReading.objects.get()
@@ -90,6 +94,9 @@ class TestDatalogger(InterceptStdoutMixin, TestCase):
             datetime(2015, 11, 10, 18, 0, 0, tzinfo=pytz.UTC)
         )
         self.assertEqual(reading.extra_device_delivered, Decimal('845.206'))
+        self.assertEqual(reading.phase_currently_delivered_l1, Decimal('0.123'))
+        self.assertEqual(reading.phase_currently_delivered_l2, Decimal('0.456'))
+        self.assertEqual(reading.phase_currently_delivered_l3, Decimal('0.789'))
 
         # Different data source.
         meter_statistics = MeterStatistics.get_solo()