+++ dsmr_frontend/models/settings.py
@@ -37,6 +37,21 @@ class FrontendSettings(SingletonModel):
         verbose_name=_('Gas delivered color'),
         help_text=_("Graph color for gas delivered")
     )
+    phase_delivered_l1_color = ColorField(
+        default='#A47448',
+        verbose_name=_('Phase L1 color'),
+        help_text=_("Graph color for phase L1")
+    )
+    phase_delivered_l2_color = ColorField(
+        default='#A4484E',
+        verbose_name=_('Phase L2 color'),
+        help_text=_("Graph color for phase L2 (when available)")
+    )
+    phase_delivered_l3_color = ColorField(
+        default='#A44882',
+        verbose_name=_('Phase L3 color'),
+        help_text=_("Graph color for phase L3 (when available)")
+    )
     temperature_color = ColorField(
         default='#0073B7',
         verbose_name=_('Temperature color'),