+++ dsmr_frontend/tests/models/test_settings.py
@@ -34,5 +34,14 @@ class TestSettings(TestCase):
     def test_gas_delivered_color(self):
         self.assertEqual(self.instance.gas_delivered_color, '#FF851B')
 
+    def test_phase_delivered_l1_color(self):
+        self.assertEqual(self.instance.phase_delivered_l1_color, '#A47448')
+
+    def test_phase_delivered_l2_color(self):
+        self.assertEqual(self.instance.phase_delivered_l2_color, '#A4484E')
+
+    def test_phase_delivered_l3_color(self):
+        self.assertEqual(self.instance.phase_delivered_l3_color, '#A44882')
+
     def test_temperature_color(self):
         self.assertEqual(self.instance.temperature_color, '#0073B7')