+++ test-all.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
 
 echo ""
+echo "--- Running Pylama for code audit..."
+pylama
+
+# Abort when audit fails.
+if [ $? -ne 0 ]; then
+    echo "[!] Code audit failed"
+    exit;
+fi
+
+echo "OK"
+
+echo ""
 echo "--- Testing with SQLite (4 processes)..."
 pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_sqlite -n 4
 
@@ -13,7 +25,3 @@ pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term -
 echo ""
 echo "--- Testing with MySQL (1 proces due to concurrency limitations)..."
 pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_mysql
-
-echo ""
-echo "--- Running Pylama for code audit..."
-pylama